removed drafting notebook. reverted classed_fixed.py to match classes.py. Removed unused functions from pvpg_fixed.
