+++ scripts/pvpg_fixed.py
@@ -70,7 +70,7 @@ def pvpg(atl03path, atl08path, j = None):
         #Orthogonal Distance Regression with simple initial guess
         linear = Model(linear_model)
         data = Data(atl08.df.Eg,atl08.df.Ev)
-        odr = ODR(data, linear, beta0 = [1.0,1.0])
+        odr = ODR(data, linear, beta0 = [-1.0,np.max(atl08.df.Ev)])
         result = odr.run()
         slope, intercept = result.beta
         
@@ -172,226 +172,7 @@ def pvpg_flagged(atl03path, atl08path, j = None):
     plt.tight_layout(rect=[0, 0, 1, 0.97])  # Adjust the layout to make room for the suptitle
     plt.show()
     return
-    
-def pvpg_gt(atl03path, atl08path, gt):
-    """
-    Variation of pvpg function that focuses on a single groundtrack. The purpose of this
-    function is to compare a single beam in different files. This is a quick way to reduce
-    the quantity of output (1 track vs 6 tracks shown) and retaining diversity of output.
-    Good for experimenting with methods and checking for a degree of generality.
-
-    atl03path - Path/to/ATL03/file
-    atl08path - Path/to/matching/ATL08/file
-    gt - One of "gt1l", "gt1r", "gt2l", "gt2r", "gt3l", or "gt3r", indiicating the chosen ground track to focus on.
-    """
-
-    i = 0
-
-    fig, axes = plt.subplots(1, 2, figsize=(10, 5))
-    ax = axes.flatten()
-
-    # Extracting date and time from the filename
-    title_date = parse_filename_datetime(atl03path)
-
-    # Set the figure title
-    fig.suptitle(title_date, fontsize=16)
-
-    try:
-        atl03 = ATL03(atl03path, atl08path, gt)
-    except (KeyError, ValueError, OSError) as e:
-        print('Sucks to suck.')
-    atl08 = ATL08(atl08path, gt)
-
-    atl03.plot(ax[i])
-
-    def linear_model(params, x):
-        return params[0]*x + params[1]
-
-    linear = Model(linear_model)
-    data = Data(atl08.df.Eg,atl08.df.Ev)
-    odr = ODR(data, linear, beta0 = [1.0,1.0])
-    result = odr.run()
-    slope, intercept = result.beta
-
-    ax[i+1].set_title(f"{gt} 100m Photon Rates")
-    ax[i+1].scatter(atl08.df.Eg, atl08.df.Ev, s=10)
-    ax[i+1].plot([0,-intercept/slope],[intercept,0])
-    ax[i+1].set_xlabel('Eg (returns/shot)')
-    ax[i+1].set_ylabel('Ev (returns/shot)')
-    ax[i+1].set_xlim(0,8)
-    ax[i+1].set_ylim(0,8)
-    ax[i+1].annotate(r'$\rho_v/\rho_g \approx {:.2f}$'.format(-slope),
-                   xy=(.95,.95),
-                   xycoords='axes fraction',
-                   ha='right',
-                   va='top',
-                   bbox=dict(boxstyle="round,pad=0.3",
-                             edgecolor="black",
-                             facecolor="white"))
-
-    plt.tight_layout(rect=[0, 0, 1, 0.97])  # Adjust the layout to make room for the suptitle
-    plt.show()
-    return
-    
-def pvpg_gt_penalized(atl03path, atl08path, gt, f_scale = .1, loss = 'arctan', model = model, residuals = residuals):
-    """
-    Experimentation function focused on single groundtracks to look at various regression
-    methods. In particular, the focus is on outlier robust regression.
-
-    atl03path - Path/to/ATL03/file
-    atl08path - Path/to/matching/ATL08/file
-    gt - One of "gt1l", "gt1r", "gt2l", "gt2r", "gt3l", or "gt3r", indiicating the chosen ground track to focus on.
-    f_scale - Parameter in least_squares() function when loss is nonlinear, indiciating the value of the soft margin between inlier and outlier residuals.
-    loss - string for loss parameter in least_squares()
-    """
-    
-    i = 0
-
-    try:
-        atl03 = ATL03(atl03path, atl08path, gt)
-    except (KeyError, ValueError, OSError) as e:
-        # print('Sucks to suck.')
-        return
-    atl08 = ATL08(atl08path, gt)
-
-    fig, axes = plt.subplots(1, 2, figsize=(10, 5))
-    ax = axes.flatten()
-
-    # Extracting date and time from the filename
-    title_date = parse_filename_datetime(atl03path)
-
-    # Set the figure title
-    fig.suptitle(title_date, fontsize=16)
-
-    atl03.plot(ax[i])
-
-    X = atl08.df.Eg
-    Y = atl08.df.Ev
-        
-    initial_guess = [-1,(np.median(Y) + np.max(Y))]
-    
-    
-    result = least_squares(residuals, initial_guess, loss=loss, f_scale=f_scale, args=(X, Y))
-    
-    a_opt, b_opt = result.x
-
-    ax[i+1].set_title(f"{gt} 100m Photon Rates")
-    ax[i+1].scatter(X, Y, s=10)
-    ax[i+1].plot(np.array([-10,20]), model([a_opt, b_opt], np.array([-10,20])), label='Orthogonal Distance Regression', color='red', linestyle='--')
-    ax[i+1].set_xlabel('Eg (returns/shot)')
-    ax[i+1].set_ylabel('Ev (returns/shot)')
-    ax[i+1].set_xlim(0,12)
-    ax[i+1].set_ylim(0,12)
-    ax[i+1].annotate(r'$\rho_v/\rho_g \approx {:.2f}$'.format(-a_opt),
-                   xy=(.95,.95),
-                   xycoords='axes fraction',
-                   ha='right',
-                   va='top',
-                   bbox=dict(boxstyle="round,pad=0.3",
-                             edgecolor="black",
-                             facecolor="white"))
-
-    plt.tight_layout(rect=[0, 0, 1, 0.97])  # Adjust the layout to make room for the suptitle
-    plt.show()
-    return
-    
-def pvpg_penalized(atl03path, atl08path,f_scale = [.1,.1], loss = ['arctan','arctan'], bounds0 = ([-100, 0], [-1/100, 16]), bounds1 = ([-100, 0], [-1/100, 16]), file_index = None, res0 = residuals, res1 = residuals, model = model, x_guess = [-1,-1], y_guess = [np.max,np.max]):
-    """
-    Experimentation function no longer restricted to the ground tracks. Additionally allows
-    visual comparison between two different penalized regression methods.
-
-    atl03path - Path/to/ATL03/file
-    atl08path - Path/to/matching/ATL08/file
-    f_scale - Parameter in least_squares() function when loss is nonlinear, indiciating the value of the soft margin between inlier and outlier residuals. This is an array of two values, one for each regression.
-    loss - string for loss parameter in least_squares(). Array of two values, one for each regression.
-    bounds0, bounds1 - bounds for slope of line and y-intercept in shape ([--,--],[--,--]), allowing us to restrict the regression to realistic values. One for each regression.
-    file_index - Index of file if cycling through an array of filenames, displayed in figure titles for a given file. Allows us to easily pick out strange cases for investigation.
-    res - Default holds the ODR residuals function to be used in least_squares(). Can hold adjusted residual functions as well.
-    x_guess - Initial guess for regression slope for each regression, default set to -1.
-    y_guess - Function used to produce initial guess for y_intercept, default set as highest Ev return in track.
-    """
-    
-    i = 0
-    
-    tracks = ['gt1r', 'gt1l', 'gt2r', 'gt2l', 'gt3r', 'gt3l']
-    #A = h5py.File(atl03path, 'r')
-    
-    fig, axes = plt.subplots(6, 3, figsize=(8, 30))
-    ax = axes.flatten()
-    
-    # Extracting date and time from the filename
-    title_date = parse_filename_datetime(atl03path)
-
-    # Set the figure title
-    if file_index != None:
-        fig.suptitle(title_date + ' - N = ' + str(file_index), fontsize=16)
-    else:
-        fig.suptitle(title_date, fontsize=16)
-
-    for gt in tracks:
-        
-        try:
-            atl03 = ATL03(atl03path, atl08path, gt)
-        except (KeyError, ValueError, OSError) as e:
-            i += 3
-            continue
-        atl08 = ATL08(atl08path, gt)
-
-        atl03.plot(ax[i])
-
-        X = atl08.df.Eg
-        Y = atl08.df.Ev
-        
-        initial_guess = [x_guess[0],y_guess[0](Y)]
-        
-        result = least_squares(res0, initial_guess, loss=loss[0], f_scale=f_scale[0], args=(X, Y), bounds = bounds0)
-            
-        a_guess, b_guess = result.x
-        
-        ax[i+1].set_title(f"{gt} Photon Rates", fontsize=8)
-        ax[i+1].scatter(X, Y, s=10)
-        ax[i+1].plot(np.array([-10,20]), model([a_guess, b_guess], np.array([-10,20])), label='Orthogonal Distance Regression', color='red', linestyle='--')
-        ax[i+1].set_xlabel('Eg (returns/shot)')
-        ax[i+1].set_ylabel('Ev (returns/shot)')
-        ax[i+1].set_xlim(0,12)
-        ax[i+1].set_ylim(0,12)
-        ax[i+1].annotate(r'$\rho_v/\rho_g \approx {:.2f}$'.format(-a_guess),
-                       xy=(.95,.95),
-                       xycoords='axes fraction',
-                       ha='right',
-                       va='top',
-                       bbox=dict(boxstyle="round,pad=0.3",
-                                 edgecolor="black",
-                                 facecolor="white"))
-        
-        initial_guess = [x_guess[1],y_guess[1](Y)]
-        
-        result = least_squares(res1, initial_guess, loss=loss[1], f_scale=f_scale[1], args=(X, Y), bounds = bounds1)
-        
-        a_opt, b_opt = result.x
-        
-        ax[i+2].set_title(f"{gt} Photon Rates", fontsize=8)
-        ax[i+2].scatter(X, Y, s=10)
-        ax[i+2].plot(np.array([-10,20]), model([a_opt, b_opt], np.array([-10,20])), label='Orthogonal Distance Regression', color='red', linestyle='--')
-        ax[i+2].set_xlabel('Eg (returns/shot)')
-        ax[i+2].set_ylabel('Ev (returns/shot)')
-        ax[i+2].set_xlim(0,12)
-        ax[i+2].set_ylim(0,12)
-        ax[i+2].annotate(r'$\rho_v/\rho_g \approx {:.2f}$'.format(-a_opt),
-                       xy=(.95,.95),
-                       xycoords='axes fraction',
-                       ha='right',
-                       va='top',
-                       bbox=dict(boxstyle="round,pad=0.3",
-                                 edgecolor="black",
-                                 facecolor="white"))
-        i += 3
-    
-    plt.tight_layout(rect=[0, 0, 1, 0.97])  # Adjust the layout to make room for the suptitle
-    plt.show()
-    
-    plt.close(fig)
-    return
+   
     
 def pvpg_penalized_flagged(atl03path, atl08path,f_scale = [.1,.1], loss = ['arctan','arctan'], bounds0 = ([-100, 0], [-1/100, 16]), bounds1 = ([-100, 0], [-1/100, 16]), file_index = None, res0 = residuals, res1 = residuals, model = model, x_guess = [-1,-1], y_guess = [np.max,np.max]):
     """
@@ -497,8 +278,6 @@ def pvpg_penalized_flagged(atl03path, atl08path,f_scale = [.1,.1], loss = ['arct
     
     plt.tight_layout(rect=[0, 0, 1, 0.97])  # Adjust the layout to make room for the suptitle
     plt.show()
-    
-    plt.close(fig)
     return
     
 def pvpg_strongweak(atl03path, atl08path, xlim, ylim, j = None):
